\name{last_pushes}
\alias{last_pushes}
\title{List recent successful pushes}
\usage{last_pushes(n = 20)}
\arguments{
\item{n}{Maximum number of pushes to return.}
}
\value{A data frame of recent successful push log entries, newest first.}
\description{List recent successful pushes from the pusher log.}
