% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{notification_style}
\alias{notification_style}
\title{Show the successful push notification style}
\usage{
notification_style()
}
\value{
Either \code{"banner"} for Notification Center banners or \code{"alert"} for
persistent macOS alerts.
}
\description{
Show the successful push notification style
}
