% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{notifications_enabled}
\alias{notifications_enabled}
\title{Check whether push notifications are enabled}
\usage{
notifications_enabled()
}
\value{
\code{TRUE} if notifications are enabled, otherwise \code{FALSE}.
}
\description{
Notifications are off by default.
}
