% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{scheduler_status}
\alias{scheduler_status}
\title{Show macOS launchd scheduler status}
\usage{
scheduler_status()
}
\value{
A data frame with scheduler status.
}
\description{
Show macOS launchd scheduler status
}
