% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_notifications}
\alias{set_notifications}
\title{Enable or disable successful push notifications}
\usage{
set_notifications(enabled = TRUE)
}
\arguments{
\item{enabled}{\code{TRUE} to enable notifications, \code{FALSE} to disable them.}
}
\value{
The updated notifications setting, invisibly.
}
\description{
Enable or disable successful push notifications
}
