% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Show pusher status for all tracked repositories}
\usage{
status()
}
\value{
A data frame with unpublished and due commit counts.
}
\description{
Show pusher status for all tracked repositories
}
