% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upcoming.R
\name{upcoming_pushes}
\alias{upcoming_pushes}
\title{List upcoming unpublished commits}
\usage{
upcoming_pushes(n = 20)
}
\arguments{
\item{n}{Maximum number of upcoming commits to return.}
}
\value{
A data frame of future unpublished commits, including commit titles,
ordered by effective date.
}
\description{
List upcoming unpublished commits
}
